crm-5: refactoring enum and trait files and move them to their own place
