+++ app/Enum/Can.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Enum;
 
 enum Can: string
 {
